Add a Logger dependency
Add a very basic Logger implementation - mostly
to prove the usage of evt-dependency.

When SlackEventReceived is instantiated with `new`
(as it is in the tests), it gets a null-object implementation
of the Logger. You will not see any log output.
When SlackEventReceived is instantiated with `build`,
the Logger implemented defined in lib/logger.rb is used.

That is the version used when deployed - and we will see log output.

Use `.build` to instantiate objects with their operational dependencies,
and `.new` to instantiate objects with their inert/null object dependencies.