Spike referencing a SNS trigger from another lambda
The SNS Topic is defined as a trigger by the 'mark'
function. The 'react' function needs to know the full
name so it can publish to the topic.

This approach works, but the topic name that serverless auto-creates
is not namespaced at all. We probably want service/stage in the topic name.