Scope lambda permissions to functions in this service
Instead of granting granting the IAM role permission to invoke any
other lambda (`"*"`), scope it to lambdas within this service.

We really only need the ability to invoke MarkLambdaFunction, but if
you try to grant that permission to the shared IAM role, the deploy
fails with a circular reference error (since the role is shared by all
lambdas, it is trying to grant permission to invoke MarkLambdaFunction
to the MarkLambdaFunction). Hence the wildcard.

To build the ARN, I need access to the AWS region and Account ID, which
are not available unless you drop to raw CloudFormation syntax (using FN:Joins).
The `serverless-pseudo-parameters` plugin lets you reference the
CloudFormation Pseudo Parameters (https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/pseudo-parameter-reference.html) using a special variable syntax.
The plugin will pre-process the yaml file and replace usages with the
raw CloudFormation syntax.